+++ RELEASES.md
@@ -1,6 +1,7 @@
 Version 0.9.1 (2022-12-XX)
 ========================
 * New driving model
+* Removed driver monitoring timer resetting on interaction if face detected and distracted
 * Kia Sorento Plug-in Hybrid 2022 support thanks to sunnyhaibin!